[OT-220][FIX]: apis 내부에 타입 선언으로 수정
