+++ src/entities/statistics/apis/getTagStatsByCategory.ts
@@ -1,6 +1,11 @@
 // 태그별 시청 통계 조회
 import { api } from "@shared/api";
-import { ApiResponse, TagStat } from "@shared/types";
+import { ApiResponse } from "@shared/types";
+
+export interface TagStat {
+  tagName: string;
+  viewCount: number;
+}
 
 export const getTagStatsByCategory = async (categoryId: number) => {
   const res = await api.get<ApiResponse<TagStat[]>>(
