+++ src/entities/statistics/apis/getTagsByCategory.ts
@@ -1,6 +1,11 @@
 // 카테고리별 태그 목록 조회
 import { api } from "@shared/api";
-import { ApiResponse, Tag } from "@shared/types";
+import { ApiResponse } from "@shared/types";
+
+export interface Tag {
+  tagId: number;
+  name: string;
+}
 
 export const getTagsByCategory = async (categoryId: number) => {
   const res = await api.get<ApiResponse<Tag[]>>(`/admin/tags/${categoryId}`);
