+++ src/shared/types/index.ts
@@ -7,8 +7,3 @@ export * from "./videoFileMeta";
 export type { ApiError } from "./apiError";
 export type { ApiResponse } from "./apiResponse";
 export type { PageInfo, BasePaginationParams } from "./pagination";
-export type {
-  IngestStatus,
-  IngestJob,
-  IngestJobListResponse,
-} from "./ingestJob";
